Integrate UserSettings to PersistenceClient module
