+++ Sources/ClientModels/UserSettings.swift
@@ -3,7 +3,7 @@ import Styleguide
 import SwiftUI
 import UIKit
 
-public struct UserSettingsV2: Codable, Equatable {
+public struct UserSettings: Codable, Equatable {
   public var appIcon: AppIcon?
   public var colorScheme: ColorScheme
   public var enableGyroMotion: Bool
@@ -57,37 +57,5 @@ public struct UserSettingsV2: Codable, Equatable {
     (try? container.decode(Bool.self, forKey: .enableReducedAnimation)) ?? false
     self.musicVolume = (try? container.decode(Float.self, forKey: .musicVolume)) ?? 1
     self.soundEffectsVolume = (try? container.decode(Float.self, forKey: .soundEffectsVolume)) ?? 1
-  }
-}
-
-public enum AppIcon: String, Codable, CaseIterable, Hashable {
-  case icon1 = "icon-1"
-  case icon2 = "icon-2"
-  case icon3 = "icon-3"
-  case icon4 = "icon-4"
-  case icon5 = "icon-5"
-  case icon6 = "icon-6"
-  case icon7 = "icon-7"
-  case icon8 = "icon-8"
-  case iso = "icon-iso"
-
-  var color: Color {
-    switch self {
-    case .icon1:
-      return .isowordsYellow
-    case .icon2:
-      return .isowordsOrange
-    case .icon3:
-      return .isowordsRed
-
-    case .icon4, .icon5, .icon6, .icon7, .icon8, .iso:
-      return Color(
-        UIColor { trait in
-          trait.userInterfaceStyle == .light
-          ? .black
-          : .white
-        }
-      )
-    }
   }
 }
