+++ Sources/PersistenceClient/Client.swift
@@ -2,8 +2,8 @@ import ClientModels
 import Foundation
 
 public struct PersistenceClient {
-  public var userSettings: () async -> Data
-  public var setUserSettings: (Data) -> Void
+  public var userSettings: () async -> UserSettings
+  public var setUserSettings: (UserSettings) async -> Void
 
   public var savedGames: () async -> SavedGamesState
   public var setSavedGames: (SavedGamesState) async -> Void
